adds a check to see if the commit exists on given branch
After a bunch of discussion & thoughts (there is no way to check if a
commit belongs to a branch using the github api, except for strolling
through all commits of said branch maybe), a solution on the script that
calls the test suite got added. This obtains a list of all remote
branches that contain the given commit, and then checks if the given
branch is in it. Probably not fully correct, but it should do for now.